LayoutTests/userscripts/script-not-run-for-fragments-expected.txt

CONSOLE MESSAGE: injected: blank.xhtml
This test should not infinitely inject user scripts.
